Solve PR's comments
