+++ telegram/src/main/kotlin/com/github/kotlintelegrambot/Bot.kt
@@ -31,7 +31,7 @@ import com.github.kotlintelegrambot.network.call
 import com.github.kotlintelegrambot.network.serialization.GsonFactory
 import com.github.kotlintelegrambot.types.DispatchableObject
 import com.github.kotlintelegrambot.types.TelegramBotResult
-import com.github.kotlintelegrambot.updater.ExecutorLooper
+import com.github.kotlintelegrambot.updater.CoroutineLooper
 import com.github.kotlintelegrambot.updater.Updater
 import com.github.kotlintelegrambot.webhook.WebhookConfig
 import com.github.kotlintelegrambot.webhook.WebhookConfigBuilder
@@ -81,7 +81,7 @@ class Bot private constructor(
 
         fun build(): Bot {
             val updatesQueue = LinkedBlockingQueue<DispatchableObject>()
-            val looper = ExecutorLooper(Dispatchers.IO)
+            val looper = CoroutineLooper(Dispatchers.IO)
             val apiClient = ApiClient(token, apiUrl, timeout, logLevel, proxy, gson)
             val updater = Updater(looper, updatesQueue, apiClient, timeout)
             val dispatcher = Dispatcher(
