+++ telegram/src/main/kotlin/com/github/kotlintelegrambot/updater/Looper.kt
@@ -5,7 +5,6 @@ import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.yield
-import java.util.concurrent.Executor
 
 internal interface Looper {
     fun loop(loopBody: suspend () -> Unit)
@@ -13,16 +12,17 @@ internal interface Looper {
 }
 
 /**
- * [Looper] implementation that runs a given block of code in a loop with an [Executor] (mostly
- * intended to run the loop in a different thread). The loop will stop if the thread running the
+ * [Looper] implementation that runs a given block of code in a loop with an [CoroutineDispatcher] (mostly
+ * intended to run the loop in a different coroutine). The loop will stop if the coroutine running the
  * loop is interrupted or in the next iteration after the [quit] method is called.
  */
-internal class ExecutorLooper(ioDispatcher: CoroutineDispatcher) : Looper {
+internal class CoroutineLooper(ioDispatcher: CoroutineDispatcher) : Looper {
 
     private val scope: CoroutineScope = CoroutineScope(ioDispatcher)
     private var job: Job? = null
 
     override fun loop(loopBody: suspend () -> Unit) {
+        job?.cancel()
         job = scope.launch { runLoop(loopBody) }
     }
